scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
googlemaps==4.10.0
python-dotenv==1.0.1
rapidfuzz==3.9.3
playwright==1.49.1
geopy==2.4.1
curl_cffi>=0.14.0
